handle no lines selected
