+++ client/packages/inventory/src/Stocktake/api/hooks/utils/index.ts
@@ -1,7 +1,9 @@
 import { useIsStocktakeDisabled } from './useIsStocktakeDisabled';
+import { useSelectedRows } from './useSelectedRows';
 import { useStocktakeApi } from './useStocktakeApi';
 
 export const Utils = {
   useIsStocktakeDisabled,
   useStocktakeApi,
+  useSelectedRows,
 };
